+++ partner-chains-cli/src/prepare_configuration/prepare_main_chain_config.rs
@@ -6,10 +6,10 @@ use crate::config::{
 	get_cardano_network_from_file, CardanoNetwork, SidechainParams, PC_CONTRACTS_CLI_PATH,
 };
 use crate::io::IOContext;
-use crate::prepare_configuration::prepare_cardano_params::prepare_cardano_params;
 use crate::pc_contracts_cli_resources::{
 	establish_pc_contracts_cli_configuration, PcContractsCliResources,
 };
+use crate::prepare_configuration::prepare_cardano_params::prepare_cardano_params;
 use crate::smart_contracts;
 use anyhow::anyhow;
 use serde_json::Value;