+++ partner-chains-cli/src/setup_main_chain_state/tests.rs
@@ -1,12 +1,12 @@
 use crate::config::config_fields::CARDANO_PAYMENT_SIGNING_KEY_FILE;
 use crate::config::{config_fields, PC_CONTRACTS_CLI_PATH};
 use crate::config::{CHAIN_CONFIG_FILE_PATH, RESOURCES_CONFIG_FILE_PATH};
+use crate::pc_contracts_cli_resources::tests::establish_pc_contracts_cli_configuration_io;
+use crate::pc_contracts_cli_resources::PcContractsCliResources;
 use crate::prepare_configuration::tests::{
 	prompt_and_save_to_existing_file, prompt_with_default_and_save_to_existing_file,
 };
 use crate::setup_main_chain_state::SetupMainChainStateCmd;
-use crate::pc_contracts_cli_resources::tests::establish_pc_contracts_cli_configuration_io;
-use crate::pc_contracts_cli_resources::PcContractsCliResources;
 use crate::tests::{MockIO, MockIOContext};
 use crate::CmdRun;
 use serde_json::json;